Fix errors and simplify code
A new `NoValidTokenException` is added and raised if a combination of valid characters does not together form a valid token.

The `advance()` instance method is moved from `get_next_token()` to `tokenize()` so that it can be called only once.

Small errors fixed and a docstring added to the `number()` method since it's recursive